Be less invasive to the post-update hook

+++ githooks/update.rb
@@ -7,14 +7,18 @@ CONFIG = DeployConfig.new(ARGV)
 
 exit(0) unless CONFIG.deployable?
 post_update = File.join(CONFIG.repo, 'hooks', 'post-update')
-puts '*' * 100
-puts post_update
-puts '*' * 100
-File.open(post_update, File::CREAT|File::WRONLY) do |f|
-  f.puts('#!/bin/bash')
-  f.puts("#{File.dirname(File.absolute_path(__FILE__))}/deploy.rb #{CONFIG.ref} #{CONFIG.old} #{CONFIG.new} #{CONFIG.repo}")
-  f.puts('rm $0')
+if File.exist?(post_update)
+  File.open(post_update, File::APPEND|File::WRONLY) do |f|
+    f.puts("#{File.dirname(File.absolute_path(__FILE__))}/deploy.rb #{CONFIG.ref} #{CONFIG.old} #{CONFIG.new} #{CONFIG.repo}")
+    f.puts("sed -i '/#{CONFIG.old} #{CONFIG.new}/d' #{post_update}")
+  end
+else
+  File.open(post_update, File::CREAT|File::WRONLY) do |f|
+    f.puts('#!/bin/bash')
+    f.puts("#{File.dirname(File.absolute_path(__FILE__))}/deploy.rb #{CONFIG.ref} #{CONFIG.old} #{CONFIG.new} #{CONFIG.repo}")
+    f.puts('rm $0')
+  end
+  File.chmod(0755, post_update)
 end
-File.chmod(0755, post_update)
